+++ memberaudit/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "memberaudit.apps.MemberAuditConfig"
 
-__version__ = "2.4.2"
+__version__ = "2.5.0a3"
 __title__ = "Member Audit"