+++ CHANGELOG.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+## [2.5.0] - TBD
+
+### Added
+
+- General
+  - Automatically disable characters which no longer have an owner
+  - Disabled characters are re-enabled when user adds them again
+
+- Admin page for characters
+  - Ability to manually disable (and re-enable) sync for characters
+  - New column "Update Status" for characters on the admin page giving better information:
+    - OK: Update complete and successful
+    - In Progress: Update has not completed yet and is most likely ongoing
+    - Incomplete: Some sections are missing completely. This is most likely an issue.
+    - Error: An error occurred during the update
+    - Disabled: Sync for this character was disabled
+  - New filter with counts for update status
+  - "No Main" added to State filter and also now showing counts
+
 ## [2.4.2] - 2023-01-31
 
 ### Changed